Separate manual server-client
